backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sentence-transformers>=3.0.0
chromadb>=0.6.0
Pillow>=10.4.0
python-multipart>=0.0.9
numpy>=2.0.0
